# docs/openapi.yaml
components:
  schemas:
    ApiAgentAnswerDTO:
      properties:
        agentic:
          type: boolean
        answer:
          type: string
        question:
          type: string
        stepBudgetExhausted:
          type: boolean
        steps:
          items:
            $ref: '#/components/schemas/ApiAgentStepDTO'
          nullable: true
          type: array
      type: object
...
    ApiAnswerDTO:
      properties:
        answer:
          type: string
        question:
          type: string
        retrieval:
          $ref: '#/components/schemas/ApiRetrievalDTO'
      type: object
...
    ApiCreateLinkReq:
      properties:
        from:
          type: string
        note:
          type: string
        to:
          type: string
        type:
          type: string
      type: object
...
    ApiErrorResponse:
      properties:
        error:
          type: string
      type: object
    ApiGraphDTO:
      properties:
        edges:
          items:
            $ref: '#/components/schemas/ApiEdgeDTO'
          nullable: true
          type: array
        nodes:
          items:
            $ref: '#/components/schemas/ApiNodeDTO'
          nullable: true
          type: array
      type: object
...
    ApiLinkResponse:
      properties:
        from:
          type: string
        note:
          type: string
        to:
          type: string
        type:
          type: string
      type: object
...
    ApiRagAgentMessage:
      properties:
        content:
          type: string
        role:
          type: string
      type: object
    ApiRagAgentReq:
      properties:
        history:
          items:
            $ref: '#/components/schemas/ApiRagAgentMessage'
          type: array
        model:
          type: string
        question:
          type: string
      type: object
...
    ApiRagQuestionReq:
      properties:
        edgeTypes:
          items:
            type: string
          nullable: true
          type: array
        hops:
          nullable: true
          type: integer
        model:
          type: string
        question:
          type: string
        topK:
          type: integer
      type: object
...
    ApiSchemaDTO:
      properties:
        schema:
          type: string
      type: object
...
    ApiUpdateLinkReq:
      properties:
        from:
          type: string
        note:
          type: string
        to:
          type: string
        type:
          type: string
      type: object
...
    ProjectorChatModelList:
      properties:
        default:
          type: string
        models:
          items:
            type: string
          nullable: true
          type: array
      type: object
...
info:
  description: 'HTTP API for Astron: read the projected Kubernetes resource graph,
    run GraphRAG retrieval, manage manual links and saved views.'
  title: Astron API
  version: v1alpha1
openapi: 3.0.3
paths:
...
  /api/projections/{namespace}/{name}/graph:
    get:
      operationId: getGraph
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiGraphDTO'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
      summary: Get a projection's full node/edge graph
      tags:
      - projections
  /api/projections/{namespace}/{name}/links:
    delete:
      operationId: deleteLink
      parameters:
      - description: Source node id
        in: query
        name: from
        required: true
        schema:
          description: Source node id
          type: string
      - description: Target node id
        in: query
        name: to
        required: true
        schema:
          description: Target node id
          type: string
      - description: Relationship type (defaults to the manual-link type)
        in: query
        name: type
        schema:
          description: Relationship type (defaults to the manual-link type)
          type: string
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Delete a user-defined edge between two nodes
      tags:
      - links
    patch:
      operationId: updateLink
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiUpdateLinkReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiLinkResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Set or clear the note on a user-defined edge
      tags:
      - links
    post:
      operationId: createLink
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiCreateLinkReq'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiLinkResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Create a user-defined edge between two nodes
      tags:
      - links
  /api/projections/{namespace}/{name}/rag/agent:
    post:
      operationId: ragAgent
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiRagAgentReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiAgentAnswerDTO'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Answer a question with a bounded, tool-using chat agent
      tags:
      - graphrag
  /api/projections/{namespace}/{name}/rag/answer:
    post:
      operationId: ragAnswer
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiRagQuestionReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiAnswerDTO'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Retrieval-augmented answer to a natural-language question
      tags:
      - graphrag
  /api/projections/{namespace}/{name}/rag/models:
    get:
      operationId: ragModels
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectorChatModelList'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: List the chat models selectable for a projection
      tags:
      - graphrag
  /api/projections/{namespace}/{name}/rag/neighborhood:
    post:
      operationId: ragNeighborhood
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiRagNeighborhoodReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiRetrievalDTO'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: 'Structural retrieval: subgraph around a named resource'
      tags:
      - graphrag
  /api/projections/{namespace}/{name}/rag/query:
    post:
      operationId: ragQuery
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiRagQuestionReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectorQueryResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: 'Text-to-Cypher: answer a question with a guarded read-only query'
      tags:
      - graphrag
  /api/projections/{namespace}/{name}/rag/schema:
    get:
      operationId: ragSchema
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSchemaDTO'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Summarize a projection's current graph schema (kinds and relationship
        types)
      tags:
      - graphrag
  /api/projections/{namespace}/{name}/rag/search:
    post:
      operationId: ragSearch
      parameters:
      - description: Projection namespace
        in: path
        name: namespace
        required: true
        schema:
          description: Projection namespace
          type: string
      - description: Projection name
        in: path
        name: name
        required: true
        schema:
          description: Projection name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiRagSearchReq'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiRetrievalDTO'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Service Unavailable
      summary: Hybrid (vector + graph) retrieval for a query
      tags:
      - graphrag
...